% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restcountries_api.R
\name{rc_all}
\alias{rc_all}
\title{Search by All
The \code{rc_all} function described the REST endpoint available that you can use to search for all the countries information.}
\usage{
rc_all()
}
\value{

}
\description{
Search by All
The \code{rc_all} function described the REST endpoint available that you can use to search for all the countries information.
}
\examples{
\dontrun{
x <- rc_all()
x
}
}
