% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_by_blocs}
\alias{rc_by_blocs}
\title{Regional Blocs}
\usage{
rc_by_blocs(bloc)
}
\arguments{
\item{bloc}{Vector(Character) which represent the abbreviation of the regional bloc of choice eg AU represents African Union}
}
\value{

}
\description{
Search by regional blocs. A regional blocs is a union or group of countries within a specific geographical boundary. This Union or group protects its
member nations within that region from imports from the non-members. \url{https://en.wikipedia.org/wiki/Trade_bloc}
}
