% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_by_capital}
\alias{rc_by_capital}
\title{Search by Capital City}
\usage{
rc_by_capital(city)
}
\arguments{
\item{city}{Vector(character) of the capital city of the country you will like to search for.}
}
\value{

}
\description{
\code{rc_by_capital()} is used to search the restcountries restful API by the name of the city or town exercising primary status in a country, state, province
or other administrative region, usually as its seat of government. Find more information \url{https://en.wikipedia.org/wiki/Capital_city}
}
\examples{
\dontrun{
x <- rc_by_capital('Abuja')
x
}
}
