% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_by_code}
\alias{rc_by_code}
\title{Country Code}
\usage{
rc_by_code(code)
}
\arguments{
\item{code}{Vector(character) which contains the country code to query the restcountries API}
}
\value{

}
\description{
Search by ISO 3166-1 2-letter or 3-letter country code \url{https://www.iban.com/country-codes}
}
