% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restcountries_api.R
\name{rc_by_currency}
\alias{rc_by_currency}
\title{Search by Currency}
\usage{
rc_by_currency(input)
}
\arguments{
\item{input}{Vector(charact) which contains three-character alphabetic, and the three digit numeric ISO 4217 code for any country.}
}
\value{

}
\description{
Currency is a system of money in general use in a particular country at a specific time. You can read more about the currencies for the various countries \url{https://www.xe.com/currency/}
}
\examples{
\dontrun{
x <- rc_by_currency('USD')
x
}
}
