% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_by_lang}
\alias{rc_by_lang}
\title{Search by Country Language}
\usage{
rc_by_lang(lang)
}
\arguments{
\item{lang}{language symbol (string such as "en", "eng", "fr", "es")}
}
\value{

}
\description{
The \code{rc_by_lang} function returns the countries information for all the different language symbol.
}
\examples{
\dontrun{
## Countries that speaks English as either an first language or second language
x <- rc_by_lang("en")
x
}
}
