% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restcountries_api.R
\name{rc_by_name}
\alias{rc_by_name}
\title{Search by Country Full Name}
\usage{
rc_by_name(name)
}
\arguments{
\item{name}{vector(character) of the name of a country of choice.}
}
\value{

}
\description{
Search by country name. This function will allow you search countries by their short name, official name,native name or partial name.
More information can be found \url{https://en.wikipedia.org/wiki/List_of_countries_and_dependencies_and_their_capitals_in_native_languages}
}
\examples{
\dontrun{
}
}
