% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_by_region}
\alias{rc_by_region}
\title{Search by Region}
\usage{
rc_by_region(region)
}
\arguments{
\item{region}{Vector(character) of the united nations country groupings.}
}
\value{

}
\description{
Query the restcountries restful api based on united nations country grouping of world regions. You can research the grouping \url{https://www.nationsonline.org/oneworld/countries_of_the_world.htm}.
}
\examples{
\dontrun{
x = rc_by_region('africa')
x
}
}
