% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restcountries_api}
\alias{restcountries_api}
\title{Title
This function fetch the data from the RESTful API}
\usage{
restcountries_api(..., envir = parent.frame())
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title
This function fetch the data from the RESTful API
}
