% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ua}
\alias{ua}
\title{User agent for all interaction with restcountries_api}
\format{
An object of class \code{request} of length 7.
}
\usage{
ua
}
\description{
User agent for all interaction with restcountries_api
}
\keyword{datasets}
